requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.2
cbor2==5.7.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
contourpy==1.3.3
cron_descriptor==2.0.6
cryptography==46.0.2
cssselect2==0.8.0
cycler==0.12.1
daphne==4.2.1
diff-match-patch==20241021
Django==4.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-import-export==4.3.12
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.60.1
freetype-py==2.5.1
html5lib==1.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
kiwisolver==1.4.9
kombu==5.5.4
lxml==6.0.2
matplotlib==3.10.7
msgpack==1.1.2
numpy==2.3.4
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
Pillow==10.0.1
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdf==6.1.1
python-bidi==0.6.6
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tablib==3.9.0
tinycss2==1.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==8.1.1
